backend/requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
SQLAlchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.2
httpx==0.27.2
python-multipart==0.0.9
pytz==2024.2
